Merged in rotation (pull request #29)
Interpret pos-file rotation keyword

Make the PosFileReader and PosFileWriter understand the view rotation.
The view_rotation is a quaternion storing a special view rotation, that
the system should be rotated to by an external viewer. The trajectory
itself is not affected.
The PosFileReader will read the 'rotation' keyword with three Euler
angles that define the view rotation and then store it with the frame
data as a quaternion.
The PosFileWriter will translate the view rotation quaternion back to
Euler angles if not None and write the angles out behind the 'rotation'
keyword.

Approved-by: Vyas Ramasubramani <vramasub@umich.edu>
Approved-by: Carl Simon Adorf <csadorf@umich.edu>