Fix frame ordering of GetarFileReader
Frames are read from getar files using a bisection sort. The GTAR
library returns frame indices (which are strings) sorted by
(len(index), index) to support the common use case of integer frame
indices. The bisection search was using the lexicographic ordering of
strings, which caused frames to be ordered incorrectly depending on
which set of frame indices was used.

This commit augments the list of returned frames to sort properly as a
tuple rather than specifying a key to the bisect function, which is
not supported until python 3.10.