Improve the scaling of CifFileReader.
Previously, the duplicate-checking process for replicated coordinates
(after applying symmetry operations) was O(N^2) for the sake of
simplicity. This patch uses the sorting capability of `np.unique()`
for O(N logN) behavior instead.

Simultaneously, the tolerance value for finding duplicated coordinates
is now applied to the real-space coordinates, instead of fractional
coordinates, to provide a more natural criterion for elongated boxes.